refactor: split terminal jump host helpers
